+++ acc_app_optimisation/foreign_imports.py
@@ -78,20 +78,6 @@ class BackupModules:
             yield ChangeKind.ADDITION, name
 
 
-def import_all(paths: t.Iterable[str]) -> None:
-    """Call `import_from_path()` in a loop.
-
-    This is a convenience function that makes all given modules
-    available, but does not return them. It also stops on the first
-    exception and logs it.
-    """
-    try:
-        for path in paths:
-            import_from_path(path)
-    except Exception:  # pylint: disable=broad-except
-        LOG.error("aborted foreign imports", exc_info=True)
-
-
 def import_from_path(to_be_imported: str) -> ModuleType:
     """Return the module that is imported from path.